Implement reset categories
The user can press the reset toolbar button and all cells will be
assigned their original category that was automatically recognized based
on references between cells. In order to provide this functionality, the
program saves the manually set category to the CellData.SetCategory
field, so that the original category remains in CellData.Category.